+++ pkg/deployment/resources/inspector/acs.go
@@ -31,6 +31,7 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/util/globals"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/definitions"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/throttle"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/version"
 )
 
 func init() {
@@ -170,7 +171,11 @@ func (p *arangoClusterSynchronizationsInspector) Refresh(ctx context.Context) er
 	return p.state.refresh(ctx, arangoClusterSynchronizationsInspectorLoaderObj)
 }
 
-func (p arangoClusterSynchronizationsInspector) Throttle(c throttle.Components) throttle.Throttle {
+func (p *arangoClusterSynchronizationsInspector) Version() version.Version {
+	return version.V1
+}
+
+func (p *arangoClusterSynchronizationsInspector) Throttle(c throttle.Components) throttle.Throttle {
 	return c.ArangoClusterSynchronization()
 }