+++ pkg/deployment/resources/inspector/am.go
@@ -31,6 +31,7 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/util/globals"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/definitions"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/throttle"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/version"
 )
 
 func init() {
@@ -174,7 +175,11 @@ func (p *arangoMembersInspector) Refresh(ctx context.Context) error {
 	return p.state.refresh(ctx, arangoMembersInspectorLoaderObj)
 }
 
-func (p arangoMembersInspector) Throttle(c throttle.Components) throttle.Throttle {
+func (p *arangoMembersInspector) Version() version.Version {
+	return version.V1
+}
+
+func (p *arangoMembersInspector) Throttle(c throttle.Components) throttle.Throttle {
 	return c.ArangoMember()
 }