+++ pkg/deployment/resources/inspector/endpoints.go
@@ -31,6 +31,7 @@ import (
 	"github.com/arangodb/kube-arangodb/pkg/util/globals"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/definitions"
 	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/throttle"
+	"github.com/arangodb/kube-arangodb/pkg/util/k8sutil/inspector/version"
 )
 
 func init() {
@@ -170,7 +171,11 @@ func (p *endpointsInspector) Refresh(ctx context.Context) error {
 	return p.state.refresh(ctx, endpointsInspectorLoaderObj)
 }
 
-func (p endpointsInspector) Throttle(c throttle.Components) throttle.Throttle {
+func (p *endpointsInspector) Version() version.Version {
+	return version.V1
+}
+
+func (p *endpointsInspector) Throttle(c throttle.Components) throttle.Throttle {
 	return c.Endpoints()
 }